refactor LogoutButton to SignOutButton component in MainNavBar
